requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.2.0
psycopg[binary]==3.2.13
